split up promise
